test (utils): PHY class for common functions
Move PHY test harness and utilities to a separate file for sharing
across tests.
Convert to only using phytester, and cleanup the phytester code a bit.
